requirements.txt
# Core API framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# HTTP client for HuggingFace Inference API
requests==2.31.0

# Utilities
python-multipart==0.0.6
